requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
anyio==4.3.0
asgiref==3.8.1
attrs==23.2.0
certifi==2024.2.2
Django==5.0.4
frozenlist==1.4.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
multidict==6.0.5
pillow==10.3.0
python-dotenv==1.0.1
python-telegram-bot==21.1.1
sniffio==1.3.1
sqlparse==0.5.0
tzdata==2024.1
yarl==1.9.4
